Fix pyramid color bias with normalized start_point and bfloat16 inputs
Three changes to address color bias in pyramid denoising:

1. Normalize DMD start_point per-channel (zero mean, unit std) for stages > 0.
   The blended signal carries mean bias from the previous stage's x0 prediction,
   which cascades through re-noising causing monotonically growing channel means
   (-0.25 -> -0.49 -> -0.83 for ch0). Normalizing preserves spatial structure
   while removing the bias.

2. Cast latents and history to bfloat16 before model calls, matching the
   reference which uses bfloat16 throughout (model trained with bf16 activations).

3. Cast scheduler step_dmd output back to original dtype (bfloat16), matching
   the reference's convert_flow_pred_to_x0 which returns in original_dtype.

Before: R=206, G=107, B=43 (orange bias for beach prompt)
After:  R=152, G=111, B=75 (balanced warm tones)

+++ mlx_video/models/helios/scheduler.py
@@ -235,7 +235,7 @@ class HeliosScheduler:
             prev_sample = x0_pred
 
         self._step_index = cur_step + 1
-        return prev_sample
+        return prev_sample.astype(orig_dtype)
 
     def step(
         self,
